feat: 회원 탈퇴 로직 보완

+++ src/main/java/com/samdasu/dodoong/domain/party/service/PartyService.java
@@ -124,17 +124,25 @@ public class PartyService {
                         .build()
         );
 
+        party.increaseCurrentMembers();
+
         return PartyJoinResponse.from(savedPartyMember);
     }
 
     @Transactional
     public void leaveParty(Long memberId, Long partyId) {
-        findParty(partyId);
+        Party party = findPartyForUpdate(partyId);
 
         PartyMember partyMember = partyMemberRepository.findByMemberIdAndPartyId(memberId, partyId)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_JOINED_PARTY));
 
+        //파티장은 파티 탈퇴 불가(deleteParty만 가능)
+        if (partyMember.getRole() == PartyRole.LEADER) {
+            throw new CustomException(ErrorCode.PARTY_LEADER_CANNOT_LEAVE);
+        }
+
         partyMemberRepository.delete(partyMember);
+        party.decreaseCurrentMembers();
     }
 
     public PartyMonthlyMeResponse getMyMonthlyPartyStatus(
